+++ lib/domain/value_validators.dart
@@ -4,7 +4,7 @@
 // Time  : 11:28
 
 import 'package:dartz/dartz.dart';
-import 'package:kt_dart/collection.dart';
+import 'package:get_arch_core/get_arch_core.dart';
 
 import 'error/failures.dart';
 
@@ -12,6 +12,19 @@ import 'error/failures.dart';
 /// 包含常用的值验证器
 /// 在value_objects和表示层中调用,用于验证输入值是否合法
 
+///
+/// 验证值的长度在区间 [min,max] 内
+Validator_<String> vStrLength({min: 0, max: 256}) =>
+    Verify.property((str) => str.length >= 0 && str.length <= max,
+        error: ValidateError('字符串长度不在区间[$min,$max]内'));
+
+///
+/// 验证值不为null或空
+final Validator_<String> vStrNotEmpty = Verify.property(
+    (str) => str?.isNotEmpty ?? false,
+    error: ValidateError('字符串不能为null或空'));
+
+@Deprecated('请使用[validateStrLength]')
 Either<ValueFailure<String>, String> validateMaxStringLength(
   String input,
   int maxLength,
@@ -25,31 +38,11 @@ Either<ValueFailure<String>, String> validateMaxStringLength(
     ));
   }
 }
-
+@Deprecated('请使用[validateStrNotEmpty]')
 Either<ValueFailure<String>, String> validateStringNotEmpty(String input) {
   if (input.isEmpty) {
     return left(ValueFailure.empty(failedValue: input));
   } else {
     return right(input);
-  }
-}
-
-Either<ValueFailure<String>, String> validateSingleLine(String input) {
-  if (input != null && input.contains('\n')) {
-    return left(ValueFailure.multiline(failedValue: input));
-  } else {
-    return right(input);
-  }
-}
-
-Either<ValueFailure<KtList<T>>, KtList<T>> validateMaxListLength<T>(
-    KtList<T> input, int maxLength) {
-  if (input.size <= maxLength) {
-    return right(input);
-  } else {
-    return left(ValueFailure.listTooLong(
-      failedValue: input,
-      max: maxLength,
-    ));
   }
 }