+++ lib/get_arch_core.dart
@@ -4,13 +4,14 @@ export 'application/i_usecase.dart';
 
 export 'domain/error/error.dart';
 export 'domain/error/failures.dart';
-export 'domain/common_interface.dart';
+export 'domain/i_common.dart';
 export 'domain/extension.dart';
-export 'domain/value_objects.dart';
 export 'domain/value_validators.dart';
 export 'package:dartz/dartz.dart' show Either;
+export 'package:verify/verify.dart' hide ValidationError;
 
-export 'interface/i_common_interface.dart';
+
+export 'interface/i_dto.dart';
 export 'interface/i_dialog.dart';
 export 'interface/i_network.dart';
 export 'interface/i_storage.dart';