+++ lib/get_arch_other.dart
@@ -3,6 +3,10 @@
 // Date  : 2020/6/15
 // Time  : 20:58
 
+///
+/// 如非必要, 请勿使用 get_arch_other,
+/// 尽量使用 get_arch_core.dart / get_arch_part.dart
 library get_arch_other;
 
 export 'package:dartz/dartz.dart';
+export 'package:verify/verify.dart';